chore: rename logger files
