+++ electron/common/logger/abstract.logger.ts
